refactor: create components membership, topcard and settingmembership
